refactor: 병합 충돌로 인해  중복 LayoutAsset 파일 제거
